Moved auditable mixin to api-v2 package
