--- tob-api/auditable/apps.py
@@ -1,5 +1,0 @@
-from django.apps import AppConfig
-
-
-class auditableConfig(AppConfig):
-    name = 'auditable'
