+++ tob-api/tob_api/settings.py
@@ -67,7 +67,6 @@ INSTALLED_APPS = [
     "drf_generators",
     "drf_yasg",
     "django_filters",
-    "auditable",
     "api",
     "tob_api",
     "api_v2",
